+++ modules/core/artfish_overview_server.R
@@ -20,11 +20,11 @@ artfish_overview_server <- function(id, parent.session, pool, reloader){
   req(nrow(files) > 0)
   
   INFO("Get Artfish computation outputs for UI")
-  estimate <- get_artfish_results_for_ui(files, ref_fishing_units, ref_species)
+  estimate <- get_artfish_results_for_ui(input=files,input_type = "file", ref_fishing_units, ref_species)
 
   artfishr::artfish_shiny_overview_server("artfish_overview", 
                                           lang = appConfig$language, 
-                                          estimate = estimate)
+                                          estimate = reactive({ estimate }))
  
   MODULE_END_TIME <- Sys.time()
   INFO("artfish-overview: END")